+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
+## [0.6.1] - 2026-04-25
+- Added SmoothStep and MoveTowards variants for Color
+- Fix to scene controller to work correctly when the target scene is a sub-string of scene like going from "GameLevel1" to "Game" would break
+
 ## [0.6.0] - 2026-04-22
 - Added new collection StackList, a Stack that uses a List internally so you can do things like Remove and RemoveAt
 - Changes to router to use StackList and now when unregistering a route we remove remove it from the Route Stack to ensure it can't be opened if say it closes, destroys and then can't be navigated back to.
+- Routes
 
 ## [0.5.9] - 2026-04-21
 - Added SplashScreen and SplashScreenImage to GameFeatures